// backend/src/modules/aluno/dto/create-aluno.dto.ts
export class CreateAlunoDto {
  nome_completo: string;
  data_nascimento?: string;
  codigo_matricula: string;
  cpf?: string;
  rg?: string;
  orgao_expedidor?: string;
  nacionalidade: string;
  educacao_especial: boolean;
  educacao_especial_descricao?: string;
  whatsapp?: string;
  email?: string;
  endereco_rua?: string;
  endereco_numero?: string;
  endereco_bairro?: string;
  endereco_cep?: string;
  endereco_municipio?: string;
  endereco_estado?: string;
  nome_mae?: string;
  nome_pai?: string;
  contato_responsavel?: string;
  email_responsavel?: string;
  status: string;
}